fix: allow users to issue access keys to themselves
