+++ internal/server/access_keys.go
@@ -1,0 +1,53 @@
+package server
+
+import (
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/infrahq/infra/api"
+	"github.com/infrahq/infra/internal/access"
+	"github.com/infrahq/infra/internal/server/models"
+)
+
+func (a *API) ListAccessKeys(c *gin.Context, r *api.ListAccessKeysRequest) (*api.ListResponse[api.AccessKey], error) {
+	p := PaginationFromRequest(r.PaginationRequest)
+	accessKeys, err := access.ListAccessKeys(c, r.UserID, r.Name, r.ShowExpired, &p)
+	if err != nil {
+		return nil, err
+	}
+
+	result := api.NewListResponse(accessKeys, PaginationToResponse(p), func(accessKey models.AccessKey) api.AccessKey {
+		return *accessKey.ToAPI()
+	})
+
+	return result, nil
+}
+
+func (a *API) DeleteAccessKey(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
+	return nil, access.DeleteAccessKey(c, r.ID)
+}
+
+func (a *API) CreateAccessKey(c *gin.Context, r *api.CreateAccessKeyRequest) (*api.CreateAccessKeyResponse, error) {
+	accessKey := &models.AccessKey{
+		IssuedFor:         r.UserID,
+		Name:              r.Name,
+		ExpiresAt:         time.Now().UTC().Add(time.Duration(r.TTL)),
+		Extension:         time.Duration(r.ExtensionDeadline),
+		ExtensionDeadline: time.Now().UTC().Add(time.Duration(r.ExtensionDeadline)),
+	}
+
+	raw, err := access.CreateAccessKey(c, accessKey)
+	if err != nil {
+		return nil, err
+	}
+
+	return &api.CreateAccessKeyResponse{
+		ID:                accessKey.ID,
+		Created:           api.Time(accessKey.CreatedAt),
+		Name:              accessKey.Name,
+		IssuedFor:         accessKey.IssuedFor,
+		Expires:           api.Time(accessKey.ExpiresAt),
+		ExtensionDeadline: api.Time(accessKey.ExtensionDeadline),
+		AccessKey:         raw,
+	}, nil
+}
