+++ internal/server/handlers.go
@@ -71,49 +71,6 @@ type WellKnownJWKResponse struct {
 	Keys []jose.JSONWebKey `json:"keys"`
 }
 
-func (a *API) ListAccessKeys(c *gin.Context, r *api.ListAccessKeysRequest) (*api.ListResponse[api.AccessKey], error) {
-	p := PaginationFromRequest(r.PaginationRequest)
-	accessKeys, err := access.ListAccessKeys(c, r.UserID, r.Name, r.ShowExpired, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(accessKeys, PaginationToResponse(p), func(accessKey models.AccessKey) api.AccessKey {
-		return *accessKey.ToAPI()
-	})
-
-	return result, nil
-}
-
-func (a *API) DeleteAccessKey(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	return nil, access.DeleteAccessKey(c, r.ID)
-}
-
-func (a *API) CreateAccessKey(c *gin.Context, r *api.CreateAccessKeyRequest) (*api.CreateAccessKeyResponse, error) {
-	accessKey := &models.AccessKey{
-		IssuedFor:         r.UserID,
-		Name:              r.Name,
-		ExpiresAt:         time.Now().UTC().Add(time.Duration(r.TTL)),
-		Extension:         time.Duration(r.ExtensionDeadline),
-		ExtensionDeadline: time.Now().UTC().Add(time.Duration(r.ExtensionDeadline)),
-	}
-
-	raw, err := access.CreateAccessKey(c, accessKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return &api.CreateAccessKeyResponse{
-		ID:                accessKey.ID,
-		Created:           api.Time(accessKey.CreatedAt),
-		Name:              accessKey.Name,
-		IssuedFor:         accessKey.IssuedFor,
-		Expires:           api.Time(accessKey.ExpiresAt),
-		ExtensionDeadline: api.Time(accessKey.ExtensionDeadline),
-		AccessKey:         raw,
-	}, nil
-}
-
 func (a *API) Signup(c *gin.Context, r *api.SignupRequest) (*api.SignupResponse, error) {
 	if !a.server.options.EnableSignup {
 		return nil, fmt.Errorf("%w: signup is disabled", internal.ErrBadRequest)
